Remove: sendbird 삭제, chat2 경로를 chat로 변경
