+++ package.json
@@ -17,8 +17,6 @@
     "@mui/icons-material": "^5.14.7",
     "@mui/material": "^5.14.7",
     "@reduxjs/toolkit": "^1.9.5",
-    "@sendbird/chat": "^4.9.9",
-    "@sendbird/uikit-react": "^3.6.7",
     "@supabase/supabase-js": "^2.32.0",
     "@tanstack/react-query": "^4.32.6",
     "@types/aos": "^3.0.4",
