--- src/components/sandbird/modules/SendBirdChannelListHeader.tsx
@@ -1,5 +1,0 @@
-const SendBirdChannelListHeader = () => {
-  return <h2 style={{ width: '300px' }}>채널 리스트 헤더</h2>;
-};
-
-export default SendBirdChannelListHeader;
