--- src/components/sandbird/sendBirdGlobalStyle.css
@@ -1,4 +1,0 @@
-.sendbird-channel-list {
-  width: 100%;
-  border: 1px solid var(--sendbird-light-onlight-04);
-}
