+++ src/components/tutorInfoDetail/TutorInfoDetail.tsx
@@ -43,7 +43,7 @@ const TutorInfoDeatail = ({ id }: TutorDetailProps) => {
       const chatRoom = await getChatRoomWithTutor(session.user.id, tutorId);
 
       if (chatRoom.length > 0) {
-        navigate(`/chat2?room_id=${chatRoom[0].room_id}`);
+        navigate(`/chat?room_id=${chatRoom[0].room_id}`);
         return;
       }
 
@@ -51,7 +51,7 @@ const TutorInfoDeatail = ({ id }: TutorDetailProps) => {
 
       await inviteChatRoom(newRoom.room_id, tutorId);
 
-      navigate(`/chat2?room_id=${newRoom.room_id}`);
+      navigate(`/chat?room_id=${newRoom.room_id}`);
     } catch (error) {
       console.error(error);
     }
