+++ src/env.d.ts
@@ -2,7 +2,6 @@
 interface ImportMetaEnv {
   readonly VITE_SUPABASE_URL: string;
   readonly VITE_SUPABASE_ANON_KEY: string;
-  readonly VITE_SANDBIRD_APP_ID: string;
 }
 
 interface ImportMeta {
